requirements.txt
Pycrypto, using 'pip install pycryptodome'